[svn-r880] Made some changes for 64 bit file IDs

+++ pablo/HDF5record_RT.h
@@ -157,8 +157,8 @@ typedef struct {
 //======================================================================*/
 typedef struct fileRec {
 	struct fileRec *ptr;
-	int    hdfID;
-	int    PabloID;
+	long    hdfID;
+	long    PabloID;
 	char   *fileName;
 } fileRec_t;
 /*=======================================================================
